+++ src/main/java/com/itm/ecosurprise/repositories/IEntrega.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.itm.ecosurprise.models.Entrega;
 
-public interface IEntrega extends JpaRepository<Entrega, Long> {
+public interface IEntrega extends JpaRepository<Entrega, Integer> {
 }
